% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{calc_N}
\alias{calc_N}
\title{Calculate N based on predetermined block sizes}
\usage{
calc_N(blocks, levels, approx)
}
\arguments{
\item{blocks}{A vector of block sizes you want to use.}

\item{levels}{Number of treatment levels in the study.}

\item{approx}{The Approx. total number of randomization slots you'll want.}
}
\value{
A vector of 4 options of N based off the given inputs.
}
\description{
Calculate N based on predetermined block sizes
}
\examples{
calc_N(blocks=c(3,24), levels=3,approx=225)
}
